+++ src/__tests__/type-definition-source.test.ts
@@ -9,11 +9,4 @@ describe('type definition source', () => {
       expect(typedefs.length).not.toBe(0)
     })
   })
-
-  describe('getLatestTypingsVersion', () => {
-    it('fetches a version number', async () => {
-      const version = await subject.getLatestTypingsVersion('async')
-      expect(typeof version).toBe('string')
-    })
-  })
 })