+++ src/type-definition-source.ts
@@ -1,14 +1,9 @@
 import { ITypeDefinitionSource } from './types'
-import axios, { AxiosResponse } from 'axios'
-import { typed } from './util'
-import * as zlip from 'zlib'
+import axios from 'axios'
+import { unzipResponse } from './response-util'
 
 const typedefsUrl =
   'https://typespublisher.blob.core.windows.net/typespublisher/data/search-index-min.json'
-
-const npmClient = axios.create({
-  baseURL: 'https://registry.npmjs.org'
-})
 
 /**
  * Used to pull definitions.
@@ -18,41 +13,16 @@ export function createTypeDefinitionSource(): ITypeDefinitionSource {
     /**
      * Fetches available type defs.
      */
-    fetch: () => {
-      return axios
+    fetch: async () => {
+      const data = await axios
         .get(typedefsUrl, {
           responseType: 'stream'
         })
         .then(unzipResponse)
-        .then(data =>
-          data.map((d: any) => ({
-            typingsName: d.t
-          }))
-        )
-    },
 
-    /**
-     * Gets the latest version of a typings package.
-     */
-    getLatestTypingsVersion: (typingsPackageName: string) => {
-      return npmClient
-        .get(`${typed(typingsPackageName, true)}`)
-        .then(r => r.data['dist-tags'].latest)
+      return data.map((d: any) => ({
+        typingsName: d.t
+      }))
     }
   }
-}
-
-/**
- * Unzips a gzip-encoded response.
- *
- * @param response
- */
-function unzipResponse(response: AxiosResponse) {
-  return new Promise<Array<any>>(resolve => {
-    const unzip = zlip.createGunzip()
-    let json = ''
-    unzip.on('data', chunk => (json += chunk.toString()))
-    unzip.on('end', () => resolve(JSON.parse(json)))
-    response.data.pipe(unzip)
-  })
 }